add key result db query

+++ resources/migrations/20230811065533-create-members.up.sql
@@ -1,7 +1,7 @@
 -- 部门成员
 CREATE TABLE IF NOT EXISTS members(
     id bigserial NOT NULL PRIMARY KEY,
-    dept_id int not null,
-    user_id int not null,
+    dept_id bigint not null,
+    user_id bigint not null,
     created timestamp without time zone NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
